requirements.txt
poetry
setuptools
twine
click
openai==0.28
python-dotenv
termcolor
wheel
pytest
nox
nox-poetry
